Switched packet number from short to long to prevent wrap around. If we never wrap around, we never have a problem.

+++ src/main/java/Server.java
@@ -15,7 +15,7 @@ class Server {
 
     private String filename;
     private short packetSize;
-    private short numOfPackets;
+    private long numOfPackets;
     private HashSet<DataPacket> receivedPackets;
 
     Server(String[] args){
